// src/types/supabase.ts
export type Database = {
  public: {
    Tables: {
      restaurants: {
        Row: {
          address: string
          city: string
          description: string | null
          id: number
          name: string
          phone: string | null
          rating: string | null
          state: string
          website: string | null
          website_status: string | null
          zip: string | null
        }
        Insert: {
          address: string
          city: string
          description?: string | null
          id?: number
          name: string
          phone?: string | null
          rating?: string | null
          state: string
          website?: string | null
          website_status?: string | null
          zip?: string | null
        }
        Update: {
          address?: string
          city?: string
          description?: string | null
          id?: number
          name?: string
          phone?: string | null
          rating?: string | null
          state?: string
          website?: string | null
          website_status?: string | null
          zip?: string | null
        }
        Relationships: []
      }
      shows: {
        Row: {
          id: number
          name: string
          platform: string | null
          publisher: string | null
        }
        Insert: {
          id?: number
          name: string
          platform?: string | null
          publisher?: string | null
        }
        Update: {
          id?: number
          name?: string
          platform?: string | null
          publisher?: string | null
        }
        Relationships: []
      }
      shows_restaurants: {
        Row: {
          id: number
          restaurant_id: number
          show_id: number
        }
        Insert: {
          id?: number
          restaurant_id: number
          show_id: number
        }
        Update: {
          id?: number
          restaurant_id?: number
          show_id?: number
        }
        Relationships: [
          {
            foreignKeyName: "shows_restaurants_restaurant_id_fkey"
            columns: ["restaurant_id"]
            isOneToOne: false
            referencedRelation: "restaurants"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "shows_restaurants_show_id_fkey"
            columns: ["show_id"]
            isOneToOne: false
            referencedRelation: "shows"
            referencedColumns: ["id"]
          },
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type PublicSchema = Database[Extract<keyof Database, "public">]

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
        PublicSchema["Views"])
    ? (PublicSchema["Tables"] &
        PublicSchema["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R
      }
      ? R
      : never
    : never
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
    ? PublicSchema["Enums"][PublicEnumNameOrOptions]
    : never